# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Remote Chrome DevTools Protocol",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:12345",
    "basePath": "/",
    "paths": {
...
        "/{session_code}/navigateRequest": {
            "post": {
                "description": "Navigate Page",
                "tags": [
                    "Navigate"
                ],
                "summary": "Navigate Page",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.navigateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/refreshRequest": {
            "post": {
                "description": "Refresh Page",
                "tags": [
                    "Refresh"
                ],
                "summary": "Refresh Page",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/screenshot": {
            "get": {
                "description": "Screenshot",
                "tags": [
                    "Screenshot"
                ],
                "summary": "Screenshot",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/select_option": {
            "post": {
                "description": "Select",
                "tags": [
                    "Select"
                ],
                "summary": "Select",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.selectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/type": {
            "post": {
                "description": "Type value",
                "tags": [
                    "Type value"
                ],
                "summary": "Type value",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.typeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/viewport": {
            "post": {
                "description": "Window size",
                "tags": [
                    "Window size"
                ],
                "summary": "Window size",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.viewportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CP": {
            "type": "object",
            "properties": {
                "domain": {
                    "description": "Cookie domain.",
                    "type": "string"
                },
                "expires": {
                    "description": "Cookie expiration date, session cookie if not set",
                    "type": "string",
                    "format": "string",
                    "example": "2022-03-29T00:00:00Z"
                },
                "httpOnly": {
                    "description": "True if cookie is http-only.",
                    "type": "boolean"
                },
                "name": {
                    "description": "Cookie name.",
                    "type": "string"
                },
                "partitionKey": {
                    "description": "Cookie partition key. The site of the top-level URL the browser was visiting at the start of the request to the endpoint that set the cookie. If not set, the cookie will be set as not partitioned.",
                    "type": "string"
                },
                "path": {
                    "description": "Cookie path.",
                    "type": "string"
                },
                "priority": {
                    "description": "CookiePriority enum:\n* Low - low\n* Medium - medium\n* High - hight",
                    "type": "string",
                    "example": "Low"
                },
                "sameParty": {
                    "description": "True if cookie is SameParty.",
                    "type": "boolean"
                },
                "sameSite": {
                    "description": "CookieSameSite enum:\n* Strict - strict\n* Lax - lax\n* None - none",
                    "type": "string",
                    "example": "Strict"
                },
                "secure": {
                    "description": "True if cookie is secure.",
                    "type": "boolean"
                },
                "sourcePort": {
                    "description": "Cookie source port. Valid values are {-1, [1, 65535]}, -1 indicates an unspecified port. An unspecified port value allows protocol clients to emulate legacy cookie scope for the port. This is a temporary ability and it will be removed in the future.",
                    "type": "integer"
                },
                "sourceScheme": {
                    "description": "CookieSourceScheme enum:\n* Unset\n* NonSecure\n* Secure",
                    "type": "string",
                    "example": "Unset"
                },
                "url": {
                    "description": "The request-URI to associate with the setting of the cookie. This value can affect the default domain, path, source port, and source scheme values of the created cookie.",
                    "type": "string"
                },
                "value": {
                    "description": "Cookie value.",
                    "type": "string"
                }
            }
        },
...
        "main.dataResponse": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "evidence": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "main.internalError": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "inner_error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.navigateRequest": {
            "type": "object",
            "properties": {
                "cookies": {
                    "description": "Cookies to inject",
                    "type": "array",
                    "items": {
                        "format": "json",
                        "$ref": "#/definitions/main.CP"
                    }
                },
                "headers": {
                    "description": "Headers to inject",
                    "type": "object",
                    "format": "json",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "url": {
                    "description": "Url to navigate",
                    "type": "string",
                    "format": "string",
                    "example": "https://example.com"
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
        "main.refreshRequest": {
            "type": "object",
            "properties": {
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
        "main.selectRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "The data for select_by filter",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"city\"]"
                    ]
                },
                "select_by": {
                    "description": "Event type:\n* index - The option position\n* text - The option text\n* partialtext - The option that contains a text\n* regex - The option text/value matchs\n* value - The option value",
                    "type": "string",
                    "format": "string",
                    "example": "value"
                },
                "selector": {
                    "description": "XPATH Selector",
                    "type": "string",
                    "format": "string",
                    "example": "id(\"sel_options\")"
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
...
        "main.typeRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Data content with special keys like ArrowDown (\\u0301)",
                    "type": "string",
                    "format": "string",
                    "example": "test@test.com\\u0301"
                },
                "force": {
                    "description": "Force type",
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "normalized_content": {
                    "description": "Normalized is content withour special keys",
                    "type": "string",
                    "format": "string",
                    "example": "test@test.com"
                },
                "selector": {
                    "description": "XPATH Selector",
                    "type": "string",
                    "format": "string",
                    "example": "id(\"btn_login\")"
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
        "main.viewportRequest": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "width": {
                    "type": "integer"
                }
            }
        }
    }
}